chore: Refactor layerbuilder to use workers to fetch and process CMR granule date ranges
chore: Refactor UpdateProjection component to handle changes in active layers
