+++ web/js/mapUI/components/update-projection/updateProjection.js
@@ -369,6 +369,22 @@ function UpdateProjection(props) {
     reloadLayers(null, false);
   }, [renderedPalettes]);
 
+  const selectL2Layers = (layers) => layers
+    .filter((layer) => layer.id.includes('L2'))
+    .map((layer) => layer.granuleDateRanges);
+
+  const prevActiveLayers = usePrevious(activeLayers);
+
+  useEffect(() => {
+    if (!ui.selected) return;
+    const prevL2Layers = selectL2Layers(prevActiveLayers);
+    const activeL2Layers = selectL2Layers(activeLayers);
+    const needsReload = activeL2Layers.some((dateRange, i) => dateRange?.length !== prevL2Layers[i]?.length);
+    if (needsReload) {
+      reloadLayers(null, true);
+    }
+  }, [activeLayers]);
+
   return null;
 }
